refactor(AppService): consolidate AppService logic to separate modules as much as possible
